Move prelude variants to a separate test case
With the addition of `library/core/src/intrinsics/simd.rs`, the number
of candidates for initial completion exceeded the 500 items limit.

Following the advice of this warning message:

Your test might miss some lookup items, because only 500 most relevant items are guaranteed to be shown in the lookup. You can:
1. Make the prefix used for completion longer, so that there are less suggestions.
2. ...

We split completion of variants from prelude out, and only keep the
keywords. The Rust keywords are set with order="first", so they always
appear first.